updated code

+++ notebooks/neo4j_bulk_importer.py
@@ -58,7 +58,7 @@ def create_kg():
             os.remove(os.path.join(NEO4J_IMPORT, filename))
 
     # args.txt contains arguments for the neo4j_admin tool
-    os.path.exists(os.path.join(NEO4J_IMPORT, "args.txt")):
+    if os.path.exists(os.path.join(NEO4J_IMPORT, "args.txt")):
         os.remove(os.path.join(NEO4J_IMPORT, "args.txt"))
 
     # Copy data and metadata files into the import directory
